requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.15.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.5
django-appconf==1.0.5
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.4
django-forms-bootstrap==3.1.0
django-mathfilters==1.0.0
django-select2==7.10.0
django-weasyprint==0.5.4
html5lib==1.1
importlib-metadata==2.1.2
Pillow==9.2.0
pycparser==2.21
pyphen==0.12.0
pypng==0.0.21
PyQRCode==1.2.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tinycss2==1.1.1
typing-extensions==4.0.0
WeasyPrint==50
webencodings==0.5.1
zipp==3.6.0
